requirements.txt
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
anthropic==0.39.0
python-dotenv==1.0.0
PyYAML==6.0.1
python-dateutil==2.8.2
PyPDF2==3.0.1
